Refactored to pull out commands into their own files.
Added eslint config for naming conventions to remove TODOS.
Added test suite placeholders for missing tests.
