+++ src/test/suite/commands/configureProjectCommand.test.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright (c) 2023, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+import * as sinon from 'sinon';
+import { afterEach, beforeEach } from 'mocha';
+
+suite('Configure Project Command Test Suite', () => {
+    beforeEach(function () {});
+
+    afterEach(function () {
+        sinon.restore();
+    });
+
+    test.skip('TODO', async () => {
+        // TODO
+    });
+});
